Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/data-url-expected.txt

CONSOLE MESSAGE: 1
CONSOLE MESSAGE: 2
CONSOLE MESSAGE: 3

PASS Test data URL and scripts errors
